backend/requirements.txt
letta
letta-client
python-dotenv